# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fantasysports OpenApi Documentation",
        "description": "Fantasysports RESTful API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://146.59.15.26:8080/api/v1",
            "description": "Fantasysports API Server",
            "variables": {
                "schema": {
                    "enum": [
                        "https",
                        "http"
                    ],
                    "default": "https"
                }
            }
        }
    ],
    "paths": {
...
        "/auth/reset/password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset Password",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Accept"
                    },
                    {
                        "$ref": "#/components/parameters/Content-Type"
                    },
                    {
                        "$ref": "#/components/parameters/token"
                    },
                    {
                        "$ref": "#/components/parameters/email"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "password2"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "password2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "A reset email has been sent! Please check your email."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "405": {
                        "$ref": "#/components/responses/405"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/auth/email/verify/{id}/{hash}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify Email",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Accept"
                    },
                    {
                        "$ref": "#/components/parameters/Content-Type"
                    },
                    {
                        "$ref": "#/components/parameters/id"
                    },
                    {
                        "$ref": "#/components/parameters/hash"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Account verified."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "405": {
                        "$ref": "#/components/responses/405"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/auth/email/verify/resend": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify Resend",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Accept"
                    },
                    {
                        "$ref": "#/components/parameters/Content-Type"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "john@gmil.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Verification link sent"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "405": {
                        "$ref": "#/components/responses/405"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/contests/history": {
            "get": {
                "tags": [
                    "Contests"
                ],
                "summary": "Get Contests History",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Accept"
                    },
                    {
                        "$ref": "#/components/parameters/Content-Type"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ContestResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "405": {
                        "$ref": "#/components/responses/405"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/contests/lobby": {
            "get": {
                "tags": [
                    "Contests"
                ],
                "summary": "Get Contests Lobby",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Accept"
                    },
                    {
                        "$ref": "#/components/parameters/Content-Type"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ContestResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "405": {
                        "$ref": "#/components/responses/405"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/contests/types": {
            "get": {
                "tags": [
                    "Contests"
                ],
                "summary": "Get Contest Types",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Accept"
                    },
                    {
                        "$ref": "#/components/parameters/Content-Type"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ContestTypeResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "405": {
                        "$ref": "#/components/responses/405"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ContestResource": {
                "title": "ContestResource",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "example": "Test Contest"
                    },
                    "status": {
                        "type": "integer",
                        "enum": [
                            1,
                            2,
                            3,
                            4,
                            5
                        ]
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "user",
                            "template"
                        ]
                    },
                    "contestType": {
                        "type": "string",
                        "enum": [
                            "fifty-fifty",
                            "head-to-head",
                            "multiplier",
                            "wta",
                            "top-three",
                            "custom"
                        ]
                    },
                    "expectedPayout": {
                        "type": "number",
                        "format": "double",
                        "example": "34.17"
                    },
                    "isPrizeInPercents": {
                        "type": "integer"
                    },
                    "maxEntries": {
                        "type": "integer",
                        "example": "12"
                    },
                    "maxUsers": {
                        "type": "integer",
                        "example": "12"
                    },
                    "minUsers": {
                        "type": "integer",
                        "example": "1"
                    },
                    "leagueId": {
                        "type": "integer",
                        "example": "5"
                    },
                    "startDate": {
                        "type": "integer",
                        "example": "1650112441"
                    },
                    "endDate": {
                        "type": "integer",
                        "example": "1650122541"
                    },
                    "details": {
                        "type": "string"
                    },
                    "entryFee": {
                        "type": "number",
                        "format": "double",
                        "example": "11.87"
                    },
                    "salaryCap": {
                        "type": "integer",
                        "example": "7000"
                    },
                    "prizeBank": {
                        "type": "number",
                        "format": "double",
                        "example": "90.45"
                    },
                    "prizeBankType": {
                        "type": "integer",
                        "enum": [
                            1,
                            2,
                            3,
                            5
                        ]
                    },
                    "customPrizeBank": {
                        "type": "number",
                        "format": "double",
                        "example": "71.23"
                    },
                    "maxPrizeBank": {
                        "type": "number",
                        "format": "double",
                        "example": "300.99"
                    },
                    "suspended": {
                        "type": "integer",
                        "enum": [
                            0,
                            1
                        ]
                    },
                    "numEntries": {
                        "type": "integer",
                        "example": "1"
                    },
                    "numUsers": {
                        "type": "integer",
                        "example": "1"
                    },
                    "entries": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ContestUserResource"
                        }
                    }
                },
                "type": "object"
            },
            "ContestTypeResource": {
                "title": "ContestTypeResource",
                "properties": {
                    "value": {
                        "type": "string",
                        "enum": [
                            "wta",
                            "top-three",
                            "fifty-fifty",
                            "head-to-head",
                            "custom"
                        ]
                    },
                    "label": {
                        "type": "string",
                        "enum": [
                            "Featured",
                            "Top Three",
                            "50/50",
                            "H2H",
                            "Custom"
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "400": {
                "description": "Bad Request",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Bad request."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "401": {
                "description": "Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthenticated."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "403": {
                "description": "Forbidden",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Invalid signature."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "404": {
                "description": "Not Found",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Resource not found."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "405": {
                "description": "Method Not Allowed",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "The current method is not supported for this route."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "422": {
                "description": "Unprocessable entity",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "errors": {
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "500": {
                "description": "Internal Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Server Error"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
            "Accept": {
                "name": "Accept",
                "in": "header",
                "description": "Accept header",
                "schema": {
                    "type": "string",
                    "example": "application/vnd.api+json"
                }
            },
            "Content-Type": {
                "name": "Content-Type",
                "in": "header",
                "description": "Content-Type header",
                "schema": {
                    "type": "string",
                    "example": "application/vnd.api+json"
                }
            },
            "id": {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": "11"
                }
            },
            "hash": {
                "name": "hash",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "string",
                    "example": "12qs850keo23"
                }
            },
            "token": {
                "name": "token",
                "in": "query",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
            "email": {
                "name": "email",
                "in": "query",
                "required": true,
                "schema": {
                    "type": "string"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format **Bearer {{token}}**",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}